requirements.txt
asgiref==3.10.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.1
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
Django==5.2.8
django-filter==25.2
psycopg2-binary==2.9.9
